Use more descriptive types and improve some logic?

+++ src/main/kotlin/org/cascadebot/cascadebot/data/managers/LockManager.kt
@@ -1,65 +1,85 @@
 package org.cascadebot.cascadebot.data.managers
 
 
-import javassist.NotFoundException
 import net.dv8tion.jda.api.Permission
 import net.dv8tion.jda.api.entities.Guild
 import net.dv8tion.jda.api.entities.IPermissionHolder
-import net.dv8tion.jda.api.entities.ISnowflake
 import net.dv8tion.jda.api.entities.TextChannel
-import java.util.*
+import net.dv8tion.jda.api.requests.restaction.PermissionOverrideAction
+import net.dv8tion.jda.internal.utils.tuple.MutablePair
+import java.util.EnumSet
+
+enum class LockPermission {
+    ALLOW,
+    DENY,
+    NEUTRAL;
+
+    fun apply(action: PermissionOverrideAction, perm: EnumSet<Permission>) {
+        when (this) {
+            ALLOW -> action.grant(perm)
+            DENY -> action.deny(perm)
+            NEUTRAL -> action.clear(perm)
+        }
+    }
+
+}
 
 object LockManager {
-    // perm values: 0 = null/not specified, 1 = false, 2 = true
 
-    fun getPerm(channel: TextChannel, target: ISnowflake): List<Int> {
+    fun getPerm(channel: TextChannel, target: IPermissionHolder): MutablePair<LockPermission, LockPermission> {
         // [target, selfMember]
-        val perm = mutableListOf(0, 0)
-        try {
-            if (channel.getPermissionOverride(target as IPermissionHolder)?.denied?.contains(Permission.MESSAGE_WRITE)!!) perm[0] = 1
-            if (channel.getPermissionOverride(target)?.allowed?.contains(Permission.MESSAGE_WRITE)!!) perm[0] = 2
-            if (channel.getPermissionOverride(channel.guild.selfMember)?.allowed?.contains(Permission.MESSAGE_WRITE)!!) perm[1] = 1
-            if (channel.getPermissionOverride(channel.guild.selfMember)?.allowed?.contains(Permission.MESSAGE_WRITE)!!) perm[0] = 2
-        } catch (e: NullPointerException) {
-            null
+        val perm = MutablePair<LockPermission, LockPermission>(LockPermission.NEUTRAL, LockPermission.NEUTRAL)
+
+        val targetOverride = channel.getPermissionOverride(target)
+        if (targetOverride != null) {
+            if (targetOverride.allowed.contains(Permission.MESSAGE_WRITE)) perm.setLeft(LockPermission.ALLOW)
+            if (targetOverride.denied.contains(Permission.MESSAGE_WRITE)) perm.setLeft(LockPermission.DENY)
+        }
+
+        val selfMemberOverride = channel.getPermissionOverride(channel.guild.selfMember)
+        if (selfMemberOverride != null) {
+            if (selfMemberOverride.allowed.contains(Permission.MESSAGE_WRITE)) perm.setRight(LockPermission.ALLOW)
+            if (selfMemberOverride.denied.contains(Permission.MESSAGE_WRITE)) perm.setRight(LockPermission.DENY)
         }
         return perm
     }
 
-    private fun add(channel: TextChannel, target: ISnowflake) {
-        if (GuildDataManager.getGuildData(channel.guild.idLong).lockedChannels[channel.id] == null) {
-            (GuildDataManager.getGuildData(channel.guild.idLong).lockedChannels.put(channel.id, mutableMapOf(Pair(target.id, getPerm(channel, target)))))
+    private fun storePermissions(channel: TextChannel, target: IPermissionHolder) {
+        val lockedChannels = GuildDataManager.getGuildData(channel.guild.idLong).lockedChannels
+        val mutableMap = lockedChannels[channel.id]
+        if (mutableMap == null) {
+            lockedChannels[channel.id] = mutableMapOf(Pair(target.id, getPerm(channel, target)))
         } else {
-            GuildDataManager.getGuildData(channel.guild.idLong).lockedChannels[channel.id]!![target.id] = getPerm(channel, target)
+            mutableMap[target.id] = getPerm(channel, target)
         }
     }
 
 
     fun lock(channel: TextChannel, target: IPermissionHolder) {
-        add(channel, target)
+        storePermissions(channel, target)
         channel.upsertPermissionOverride(channel.guild.selfMember).grant(Permission.MESSAGE_WRITE).queue()
         channel.upsertPermissionOverride(target).deny(Permission.MESSAGE_WRITE).queue()
     }
 
     fun unlock(guild: Guild, channel: TextChannel, target: IPermissionHolder) {
         val state = GuildDataManager.getGuildData(channel.guild.idLong).lockedChannels[channel.id]?.get(target.id)
-                ?: throw NotFoundException("")
-        val empty = EnumSet.noneOf(Permission::class.java)
+                // If there is no state to restore, we can't do anything!
+                ?: return
         val perm = EnumSet.of(Permission.MESSAGE_WRITE)
 
-        val selfPermission = channel.getPermissionOverride(guild.selfMember)?.manager
-                ?.grant(if (state[1] == 2) perm else empty)
-                ?.deny(if (state[1] == 1) perm else empty)
-        if (state[1] == 0) selfPermission?.clear(perm)
-        selfPermission?.queue { if (it.allowedRaw == 0L && it.deniedRaw == 0L) it.delete().queue() }
+        val selfPermissionAction = channel.getPermissionOverride(guild.selfMember)?.manager
+        if (selfPermissionAction != null) {
+            state.right.apply(selfPermissionAction, perm)
+            selfPermissionAction.queue { if (it.allowedRaw == 0L && it.deniedRaw == 0L) it.delete().queue() }
+        }
 
-        val targetPermission = channel.getPermissionOverride(target)?.manager
-        if (state[0] == 2) targetPermission?.grant(perm)
-        if (state[0] == 1) targetPermission?.deny(perm)
-        if (state[0] == 0) targetPermission?.clear(perm)
-        targetPermission?.queue {
-            GuildDataManager.getGuildData(guild.idLong).lockedChannels[channel.idLong.toString()]?.remove(target.idLong.toString())
-            if (it.allowedRaw == 0L && it.deniedRaw == 0L) it.delete().queue()
+        val targetPermissionAction = channel.getPermissionOverride(target)?.manager
+        if (targetPermissionAction != null) {
+            state.left.apply(targetPermissionAction, perm)
+            targetPermissionAction.queue {
+                GuildDataManager.getGuildData(guild.idLong).lockedChannels[channel.idLong.toString()]?.remove(target.idLong.toString())
+                if (it.allowedRaw == 0L && it.deniedRaw == 0L) it.delete().queue()
+            }
         }
     }
